Fix 40 compiler errors: type mismatches, unresolved references, API incompatibilities
Fixes:
1. MockEcuTransport.kt:182 — block[12] = 142 → 142.toByte() (Int to Byte)
2. SensorCalibrationScreen — add setCalibrationType() and generateCalibration() aliases to ViewModel
3. SensorCalibrationViewModel — fix ReferenceTable field access (xBinsSize instead of size, remove dataType/scale/translate/units)
4. ConnectionViewModel — replace UsbManager.ACTION_USB_PERMISSION with hardcoded constant, fix Intent().apply{} to Intent() + setPackage()
5. CurveEditorViewModel — add undo/redo/canUndo/canRedo/setPointValue methods, fix setPointByDrag to accept Offset+Rect
6. DatalogChartScreen — fix drawLayer import (use graphicsLayerDrawLayer), fix isPressed → pressed
7. LuaConsoleScreen — explicit material3 imports instead of wildcard
8. LuaConsoleViewModel — fix LuaResult handling (data class, not sealed), fix EcuCommandCallback return type

BUILD-UNVERIFIED: Statically reviewed, awaiting build result.

+++ app/src/main/java/com/ztune/libretune/core/ecu/MockEcuTransport.kt
@@ -179,7 +179,7 @@ class MockEcuTransport(
         block[10] = (map and 0xFF).toByte()
         block[11] = ((map ushr 8) and 0xFF).toByte()
         // batteryVoltage (U08 @ offset 12) — scale 0.1, so raw = 142 → 14.2V
-        block[12] = 142
+        block[12] = 142.toByte()
         // afr (U16 @ offset 14) — scale 0.1, so raw = 147 → 14.7 AFR
         val afrRaw = 147
         block[14] = (afrRaw and 0xFF).toByte()
